Move app-shell layout constants from nav to metabase/ui
APP_BAR_HEIGHT, APP_SUBHEADER_HEIGHT, ADMIN_NAVBAR_HEIGHT, and
NAV_SIDEBAR_WIDTH are pure display-level layout dimensions consumed
well outside the nav module (App shell, detail-view, BulkActionBar).
Move them to frontend/src/metabase/ui/constants/layout.ts, export
them through the metabase/ui barrel, update all 10 importers, and
delete frontend/src/metabase/nav/constants.ts.
