+++ frontend/src/metabase/nav/components/search/SearchResultsDropdown/SearchResultsDropdown.tsx
@@ -3,8 +3,7 @@ import { jt, t } from "ttag";
 
 import type { SearchResultsFooter } from "metabase/nav/components/search/SearchResults";
 import { SearchResults } from "metabase/nav/components/search/SearchResults";
-import { APP_BAR_HEIGHT } from "metabase/nav/constants";
-import { Group, Icon, Paper, Text, rem } from "metabase/ui";
+import { APP_BAR_HEIGHT, Group, Icon, Paper, Text, rem } from "metabase/ui";
 import type { SearchContext, SearchResult } from "metabase-types/api";
 
 import S from "./SearchResultsDropdown.module.css";
