+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbar.styled.tsx
@@ -1,12 +1,11 @@
 // eslint-disable-next-line no-restricted-imports
 import styled from "@emotion/styled";
 
-import { NAV_SIDEBAR_WIDTH } from "metabase/nav/constants";
 import {
   breakpointMaxSmall,
   breakpointMinSmall,
 } from "metabase/styled-components/theme";
-import { Box, type BoxProps } from "metabase/ui";
+import { Box, type BoxProps, NAV_SIDEBAR_WIDTH } from "metabase/ui";
 
 import { SidebarLink } from "./SidebarItems";
 import { ExpandToggleButton } from "./SidebarItems/SidebarItems.styled";
