+++ frontend/src/metabase/nav/containers/MainNavbar/SidebarItems/SidebarItems.styled.tsx
@@ -7,9 +7,8 @@ import { forwardRef } from "react";
 
 import { Link } from "metabase/common/components/Link";
 import { TreeNode } from "metabase/common/components/tree/TreeNode";
-import { NAV_SIDEBAR_WIDTH } from "metabase/nav/constants";
 import type { IconProps } from "metabase/ui";
-import { Icon, Tooltip } from "metabase/ui";
+import { Icon, NAV_SIDEBAR_WIDTH, Tooltip } from "metabase/ui";
 import { alpha } from "metabase/ui/colors";
 import type { ColorName } from "metabase/ui/colors/types";
 import { color } from "metabase/ui/utils/colors";
