+++ frontend/src/metabase/ui/constants/layout.ts
